Created Character Class and setup Main
The character class has been setup with some fields and methods that
we're going to need. I know we'll need a lot more. So far you can add
story elements that correspond to turns to each character.
The main class has been setup with some pointers for each charcter and
just some code for the meantime to show how things work.